pyproject.toml
[project]
name = "antigravity-best-practices"
version = "0.1.0"
description = "FastAPI project with modern best practices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.136.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
    "httpx>=0.28.1",
    "google-adk>=1.31.1",
]
